refactor : favorite 패키지 위치 이동 (#66)
