+++ breaditnow-customer-api/src/main/java/com/breaditnow/customer/domain/bakery/service/BakeryFavoritePageService.java
@@ -1,12 +1,12 @@
-package com.breaditnow.customer.domain.bakeryfavorite.service;
+package com.breaditnow.customer.domain.bakery.service;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.breaditnow.customer.domain.bakeryfavorite.controller.req.GeoPointRequest;
-import com.breaditnow.customer.domain.bakeryfavorite.controller.res.BakeryFavoritePageResponse;
+import com.breaditnow.customer.domain.bakery.controller.req.GeoPointRequest;
+import com.breaditnow.customer.domain.bakery.controller.res.BakeryFavoritePageResponse;
 import com.breaditnow.domain.domain.favorite.dto.BakeryFavoriteDto;
 import com.breaditnow.domain.domain.favorite.repository.customerbakeryfavorite.CustomerBakeryFavoriteRepository;
 import com.breaditnow.domain.domain.vo.GeoPoint;
