+++ breaditnow-customer-api/src/main/java/com/breaditnow/customer/domain/product/controller/ProductController.java
@@ -1,0 +1,13 @@
+package com.breaditnow.customer.domain.product.controller;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import lombok.RequiredArgsConstructor;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/v1/product")
+public class ProductController {
+	
+}
