+++ breaditnow-customer-api/src/main/java/com/breaditnow/customer/domain/product/controller/ProductFavoriteController.java
@@ -1,4 +1,4 @@
-package com.breaditnow.customer.domain.productfavorite.controller;
+package com.breaditnow.customer.domain.product.controller;
 
 import java.util.Map;
 
@@ -11,10 +11,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.breaditnow.common.response.ApiSuccessResponse;
-import com.breaditnow.customer.domain.bakeryfavorite.controller.req.GeoPointRequest;
-import com.breaditnow.customer.domain.productfavorite.controller.res.ProductFavoritePageResponse;
-import com.breaditnow.customer.domain.productfavorite.service.ProductFavoritePageService;
-import com.breaditnow.customer.domain.productfavorite.service.ProductFavoriteService;
+import com.breaditnow.customer.domain.bakery.controller.req.GeoPointRequest;
+import com.breaditnow.customer.domain.product.controller.res.ProductFavoritePageResponse;
+import com.breaditnow.customer.domain.product.service.ProductFavoritePageService;
+import com.breaditnow.customer.domain.product.service.ProductFavoriteService;
 import com.breaditnow.customer.global.security.annotation.AuthCustomer;
 
 import jakarta.validation.Valid;
