+++ breaditnow-customer-api/src/main/java/com/breaditnow/customer/domain/product/controller/res/ProductFavoriteResponse.java
@@ -1,4 +1,4 @@
-package com.breaditnow.customer.domain.productfavorite.controller.res;
+package com.breaditnow.customer.domain.product.controller.res;
 
 import java.util.Arrays;
 import java.util.List;
