+++ breaditnow-customer-api/src/main/java/com/breaditnow/customer/domain/product/service/ProductFavoriteService.java
@@ -1,4 +1,4 @@
-package com.breaditnow.customer.domain.productfavorite.service;
+package com.breaditnow.customer.domain.product.service;
 
 import static com.breaditnow.domain.domain.product.enumerate.ProductType.*;
