+++ breaditnow-customer-api/src/main/java/com/breaditnow/customer/domain/product/service/ProductService.java
@@ -1,0 +1,4 @@
+package com.breaditnow.customer.domain.product.service;
+
+public class ProductService {
+}
